Code/requirements.txt
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
python-dotenv==0.20.0
scikit-learn==1.1.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
wordcloud==1.8.2.2